feat: download and extract bundles that do not have cluster resources using shell subcommand (#187)
* feat: download bundle and open shell into extracted bundle directory


* Remove commented out code


* User sbctl shell to download and interact with host bundles


* Make dir cleanup reliable


---------

+++ cli/shell.go
@@ -37,26 +37,36 @@ func ShellCmd() *cobra.Command {
 			deleteBundleDir := false
 
 			logOutput := os.Stderr
-			logFile, err := os.CreateTemp("", "sbctl-server-logs-")
+			logFile, err := os.CreateTemp("", "sbctl-server-*.log")
 			if err == nil {
-				fmt.Printf("API server logs will be written to %s\n", logFile.Name())
 				defer logFile.Close()
 				defer os.RemoveAll(logFile.Name())
+				fmt.Printf("SBCTL logs will be written to %s\n", logFile.Name())
 				log.SetOutput(logFile)
 				logOutput = logFile
 			}
 
-			go func() {
-				signalChan := make(chan os.Signal, 1)
-				signal.Notify(signalChan, os.Interrupt)
-				<-signalChan
+			cleanup := func() {
 				if kubeConfig != "" {
 					_ = os.RemoveAll(kubeConfig)
 				}
 				if deleteBundleDir && bundleDir != "" {
 					_ = os.RemoveAll(bundleDir)
 				}
+			}
+
+			go func() {
+				signalChan := make(chan os.Signal, 1)
+				// Handle Ctl-D to exit from shell
+				signal.Notify(signalChan, os.Interrupt)
+				<-signalChan
+				cleanup()
 				os.Exit(0)
+			}()
+
+			defer func() {
+				// exit from shell using "exit" command
+				cleanup()
 			}()
 
 			v := viper.GetViper()
@@ -111,47 +121,18 @@ func ShellCmd() *cobra.Command {
 				return errors.Wrap(err, "failed to find cluster data")
 			}
 
+			// If we did not find cluster data, just don't start the API server
+			if clusterData.ClusterResourcesDir == "" {
+				fmt.Println("No cluster resources found in bundle")
+				fmt.Println("Starting new shell in downloaded bundle. Press Ctl-D when done to exit from the shell")
+				return startShellAndWait(fmt.Sprintf("cd %s", bundleDir))
+			}
+
 			kubeConfig, err = api.StartAPIServer(clusterData, logOutput)
 			if err != nil {
 				return errors.Wrap(err, "failed to create api server")
 			}
 			defer os.RemoveAll(kubeConfig)
-
-			shellCmd := os.Getenv("SHELL")
-			if shellCmd == "" {
-				return errors.New("SHELL environment is required for shell command")
-			}
-
-			shellExec := exec.Command(shellCmd)
-			shellExec.Env = os.Environ()
-			fmt.Printf("Starting new shell with KUBECONFIG. Press Ctl-D when done to end the shell and the sbctl server\n")
-			shellPty, err := pty.Start(shellExec)
-			if err != nil {
-				return errors.Wrap(err, "failed to start shell")
-			}
-
-			// Handle pty size.
-			ch := make(chan os.Signal, 1)
-			signal.Notify(ch, syscall.SIGWINCH)
-			go func() {
-				for range ch {
-					if err := pty.InheritSize(os.Stdin, shellPty); err != nil {
-						log.Printf("error resizing pty: %s", err)
-					}
-				}
-			}()
-			ch <- syscall.SIGWINCH // Initial resize.
-			defer func() { signal.Stop(ch); close(ch) }()
-
-			// Set stdin to raw mode.
-			oldState, err := term.MakeRaw(syscall.Stdin)
-			if err != nil {
-				panic(err)
-			}
-			defer func() {
-				_ = term.Restore(syscall.Stdin, oldState)
-				fmt.Printf("sbctl shell exited\n")
-			}()
 
 			cmds := []string{
 				fmt.Sprintf("export KUBECONFIG=%s", kubeConfig),
@@ -159,17 +140,8 @@ func ShellCmd() *cobra.Command {
 			if v.GetBool("cd-bundle") {
 				cmds = append(cmds, fmt.Sprintf("cd %s", bundleDir))
 			}
-
-			// Setup the shell
-			setupCmd := strings.Join(cmds, "\n") + "\n"
-			_, _ = io.WriteString(shellPty, setupCmd)
-			_, _ = io.CopyN(io.Discard, shellPty, 2*int64(len(setupCmd))) // Don't print to screen, terminal will echo anyway
-
-			// Copy stdin to the pty and the pty to stdout.
-			go func() { _, _ = io.Copy(shellPty, os.Stdin) }()
-			go func() { _, _ = io.Copy(os.Stdout, shellPty) }()
-
-			return shellExec.Wait()
+			fmt.Printf("Starting new shell with KUBECONFIG. Press Ctl-D when done to exit from the shell and stop sbctl server\n")
+			return startShellAndWait(cmds...)
 		},
 	}
 
@@ -178,4 +150,52 @@ func ShellCmd() *cobra.Command {
 	cmd.Flags().Bool("cd-bundle", false, "Change directory to the support bundle path after starting the shell")
 	cmd.Flags().Bool("debug", false, "enable debug logging. This will include HTTP response bodies in logs.")
 	return cmd
+}
+
+func startShellAndWait(cmds ...string) error {
+	shellCmd := os.Getenv("SHELL")
+	if shellCmd == "" {
+		return errors.New("SHELL environment is required for shell command")
+	}
+
+	shellExec := exec.Command(shellCmd)
+	shellExec.Env = os.Environ()
+	shellPty, err := pty.Start(shellExec)
+	if err != nil {
+		return errors.Wrap(err, "failed to start shell")
+	}
+
+	// Handle pty size.
+	ch := make(chan os.Signal, 1)
+	signal.Notify(ch, syscall.SIGWINCH)
+	go func() {
+		for range ch {
+			if err := pty.InheritSize(os.Stdin, shellPty); err != nil {
+				log.Printf("error resizing pty: %s", err)
+			}
+		}
+	}()
+	ch <- syscall.SIGWINCH // Initial resize.
+	defer func() { signal.Stop(ch); close(ch) }()
+
+	// Set stdin to raw mode.
+	oldState, err := term.MakeRaw(syscall.Stdin)
+	if err != nil {
+		panic(err)
+	}
+	defer func() {
+		_ = term.Restore(syscall.Stdin, oldState)
+		fmt.Printf("sbctl shell exited\n")
+	}()
+
+	// Setup the shell
+	setupCmd := strings.Join(cmds, "\n") + "\n"
+	_, _ = io.WriteString(shellPty, setupCmd)
+	_, _ = io.CopyN(io.Discard, shellPty, 2*int64(len(setupCmd))) // Don't print to screen, terminal will echo anyway
+
+	// Copy stdin to the pty and the pty to stdout.
+	go func() { _, _ = io.Copy(shellPty, os.Stdin) }()
+	go func() { _, _ = io.Copy(os.Stdout, shellPty) }()
+
+	return shellExec.Wait()
 }
